server/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
whitenoise==6.7.0
